refactor: extract useAuth hook
