+++ src/adminPanel/pages/admin/Dashboard.tsx
@@ -2,7 +2,7 @@ import  { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContain
 import { Users, Globe, User, ShoppingBag, TrendingUp, Activity, AlertCircle, CheckCircle, Clock, Star, Trophy, Target } from 'lucide-react'; 
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../../contexts/AuthContext';
+import useAuth from '../../hooks/useAuth';
 import { useGlobalStore } from '../../store/globalStore';
 
 const Dashboard = () => {
